Update Nhap Kho

+++ LedApp/Controllers/NhapsController.cs
@@ -1,12 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using LedApp.Models;
 using Microsoft.AspNetCore.Authorization;
+using LedApp.Data;
 
 namespace LedApp.Controllers
 {
@@ -20,10 +17,25 @@ namespace LedApp.Controllers
             _context = context;
         }
 
+        // Helper dropdown TrangThai
+        private SelectList GetTrangThaiSelectList(int? selectedValue = null)
+        {
+            var list = new List<object>
+            {
+                new { Value = (int)TrangThaiNhap.DaPhanCong,  Text = "Đã phân công" },
+                new { Value = (int)TrangThaiNhap.DangBanGiao, Text = "Đang bàn giao" },
+                new { Value = (int)TrangThaiNhap.QuaThoiGian, Text = "Quá thời gian" },
+                new { Value = (int)TrangThaiNhap.HoanThanh,   Text = "Hoàn thành" }
+            };
+            return new SelectList(list, "Value", "Text", selectedValue);
+        }
+
         // GET: Nhaps
         public async Task<IActionResult> Index()
         {
-            var applicationDBContext = _context.Nhaps.Include(n => n.CuaNhap);
+            var applicationDBContext = _context.Nhaps
+                .Include(n => n.CuaNhap)
+                .Include(n => n.NhanVienXacNhan);
             return View(await applicationDBContext.ToListAsync());
         }
 
@@ -31,101 +43,92 @@ namespace LedApp.Controllers
         public async Task<IActionResult> Details(int? id)
         {
             if (id == null || _context.Nhaps == null)
-            {
                 return NotFound();
-            }
 
             var nhap = await _context.Nhaps
                 .Include(n => n.CuaNhap)
+                .Include(n => n.NhanVienXacNhan)
+                .Include(n => n.ChitietNhaps)
                 .FirstOrDefaultAsync(m => m.Id == id);
-            if (nhap == null)
-            {
-                return NotFound();
-            }
 
+            if (nhap == null) return NotFound();
             return View(nhap);
         }
 
         // GET: Nhaps/Create
         public IActionResult Create()
         {
-            ViewData["CuaNhapId"] = new SelectList(_context.CuaNhaps, "Id", "Id");
+            ViewData["CuaNhapId"]          = new SelectList(_context.CuaNhaps, "Id", "Ten");
+            ViewData["NhanVienXacNhanId"]  = new SelectList(_context.nguoiDungs, "Id", "Name");
+            ViewData["TrangThai"]          = GetTrangThaiSelectList((int)TrangThaiNhap.DaPhanCong);
             return View();
         }
 
         // POST: Nhaps/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
-        //[ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,CuaNhapId,BienSoXe,NgayNhap,GioNhap,PhutNhap,CongVao,TrangThai")] Nhap nhap)
+        public async Task<IActionResult> Create([Bind("CuaNhapId,BienSoXe,ThoiGianPhanCong,ThoiGianVaoBai,ThoiGianVaoCua,ThoiGianGioiHan,ThoiGianHoanThanh,NhanVienXacNhanId,TrangThai")] Nhap nhap)
         {
-            try 
+            try
             {
                 _context.Add(nhap);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            catch(Exception ex)
+            catch (Exception)
             {
-                ViewData["CuaNhapId"] = new SelectList(_context.CuaNhaps, "Id", "Id", nhap.CuaNhapId);
+                ViewData["CuaNhapId"]         = new SelectList(_context.CuaNhaps, "Id", "Ten", nhap.CuaNhapId);
+                ViewData["NhanVienXacNhanId"] = new SelectList(_context.nguoiDungs, "Id", "Name", nhap.NhanVienXacNhanId);
+                ViewData["TrangThai"]         = GetTrangThaiSelectList(nhap.TrangThai);
                 return View(nhap);
             }
-            
         }
 
         // GET: Nhaps/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null || _context.Nhaps == null)
-            {
-                return NotFound();
-            }
+            if (id == null || _context.Nhaps == null) return NotFound();
 
             var nhap = await _context.Nhaps.FindAsync(id);
-            if (nhap == null)
-            {
-                return NotFound();
-            }
-            ViewData["CuaNhapId"] = new SelectList(_context.CuaNhaps, "Id", "Id", nhap.CuaNhapId);
+            if (nhap == null) return NotFound();
+
+            ViewData["CuaNhapId"]         = new SelectList(_context.CuaNhaps, "Id", "Ten", nhap.CuaNhapId);
+            ViewData["NhanVienXacNhanId"] = new SelectList(_context.nguoiDungs, "Id", "Name", nhap.NhanVienXacNhanId);
+            ViewData["TrangThai"]         = GetTrangThaiSelectList(nhap.TrangThai);
             return View(nhap);
         }
 
         // POST: Nhaps/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
-        //[ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,CuaNhapId,BienSoXe,NgayNhap,GioNhap,PhutNhap,CongVao,TrangThai")] Nhap nhap)
+        public async Task<IActionResult> Edit(int id, [Bind("Id,CuaNhapId,BienSoXe,ThoiGianPhanCong,ThoiGianVaoBai,ThoiGianVaoCua,ThoiGianGioiHan,ThoiGianHoanThanh,NhanVienXacNhanId,TrangThai")] Nhap nhap)
         {
-            if (id != nhap.Id)
-            {
-                return NotFound();
-            }
+            if (id != nhap.Id) return NotFound();
+
             try
             {
-                //_context.Update(nhap);
-                //await _context.SaveChangesAsync();
-                //return RedirectToAction(nameof(Index));
-
                 var existing = await _context.Nhaps.FindAsync(id);
                 if (existing == null) return NotFound();
 
-                existing.CuaNhapId = nhap.CuaNhapId;
-                existing.BienSoXe = nhap.BienSoXe;
-                existing.NgayNhap = nhap.NgayNhap;
-                existing.GioNhap = nhap.GioNhap;
-                existing.PhutNhap = nhap.PhutNhap;
-                existing.CongVao = nhap.CongVao;
-                existing.TrangThai = nhap.TrangThai;
+                existing.CuaNhapId         = nhap.CuaNhapId;
+                existing.BienSoXe          = nhap.BienSoXe;
+                existing.ThoiGianPhanCong  = nhap.ThoiGianPhanCong;
+                existing.ThoiGianVaoBai    = nhap.ThoiGianVaoBai;
+                existing.ThoiGianVaoCua    = nhap.ThoiGianVaoCua;
+                existing.ThoiGianGioiHan   = nhap.ThoiGianGioiHan;
+                existing.ThoiGianHoanThanh = nhap.ThoiGianHoanThanh;
+                existing.NhanVienXacNhanId = nhap.NhanVienXacNhanId;
+                existing.TrangThai         = nhap.TrangThai;
 
-                _context.Entry(existing).State = EntityState.Modified; // ← thêm dòng này
+                _context.Entry(existing).State = EntityState.Modified;
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
             catch (DbUpdateConcurrencyException)
             {
-                ViewData["CuaNhapId"] = new SelectList(_context.CuaNhaps, "Id", "Id", nhap.CuaNhapId);
+                if (!NhapExists(id)) return NotFound();
+
+                ViewData["CuaNhapId"]         = new SelectList(_context.CuaNhaps, "Id", "Ten", nhap.CuaNhapId);
+                ViewData["NhanVienXacNhanId"] = new SelectList(_context.nguoiDungs, "Id", "Name", nhap.NhanVienXacNhanId);
+                ViewData["TrangThai"]         = GetTrangThaiSelectList(nhap.TrangThai);
                 return View(nhap);
             }
         }
@@ -133,44 +136,30 @@ namespace LedApp.Controllers
         // GET: Nhaps/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null || _context.Nhaps == null)
-            {
-                return NotFound();
-            }
+            if (id == null || _context.Nhaps == null) return NotFound();
 
             var nhap = await _context.Nhaps
                 .Include(n => n.CuaNhap)
+                .Include(n => n.NhanVienXacNhan)
                 .FirstOrDefaultAsync(m => m.Id == id);
-            if (nhap == null)
-            {
-                return NotFound();
-            }
 
+            if (nhap == null) return NotFound();
             return View(nhap);
         }
 
         // POST: Nhaps/Delete/5
         [HttpPost, ActionName("Delete")]
-        //[ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            if (_context.Nhaps == null)
-            {
-                return Problem("Entity set 'ApplicationDBContext.Nhaps'  is null.");
-            }
             var nhap = await _context.Nhaps.FindAsync(id);
-            if (nhap != null)
-            {
-                _context.Nhaps.Remove(nhap);
-            }
-            
+            if (nhap != null) _context.Nhaps.Remove(nhap);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
         private bool NhapExists(int id)
         {
-          return (_context.Nhaps?.Any(e => e.Id == id)).GetValueOrDefault();
+            return (_context.Nhaps?.Any(e => e.Id == id)).GetValueOrDefault();
         }
     }
 }
